REST: add proxy base class
